+++ src/cz/dubcat/xpboost/api/BoostAPI.java
@@ -6,11 +6,14 @@ import cz.dubcat.xpboost.XPBoostMain;
 import cz.dubcat.xpboost.api.MainAPI.Condition;
 import cz.dubcat.xpboost.constructors.BoostOptions;
 import cz.dubcat.xpboost.constructors.XPBoost;
+import cz.dubcat.xpboost.exceptions.BoostNotFoundException;
 
 public class BoostAPI {
 
+    /*
+     * Purchase boost for a player
+     */
     public boolean buyBoost(Player player, String boostId) {
-        
         if(!XPBoostMain.boostCfg.contains(boostId)) {
             return false;
         }
@@ -29,43 +32,12 @@ public class BoostAPI {
             }
 
             if (XPBoostMain.economy.has(player, XPBoostMain.boostCfg.getDouble(boostId + ".cost"))) {
-                int time = XPBoostMain.boostCfg.getInt(boostId + ".time");
-                double boost = XPBoostMain.boostCfg.getDouble(boostId + ".boost");
-
-                String message = XPBoostMain.getLang().getString("lang.xpbuy").replaceAll("%time%", time + "")
-                        .replaceAll("%money%", XPBoostMain.boostCfg.getString(boostId + ".cost"))
-                        .replaceAll("%boost%", String.valueOf(boost));
-                MainAPI.sendMessage(message, player);
-                XPBoostMain.economy.withdrawPlayer(player, XPBoostMain.boostCfg.getDouble(boostId + ".cost"));
-
-                XPBoost xpb = XPBoostAPI.setPlayerBoost(player.getUniqueId(), boost, time);
-
-                if (XPBoostMain.boostCfg.contains(boostId + ".behaviour")) {
-                    for (String cond : XPBoostMain.boostCfg.getConfigurationSection(boostId + ".behaviour")
-                            .getKeys(false)) {
-                        xpb.putCondition(Condition.valueOf(cond.toUpperCase()),
-                                XPBoostMain.boostCfg.getBoolean(boostId + ".behaviour." + cond));
-                    }
+                try {
+                    this.giveBoost(player, boostId, null);
+                    XPBoostMain.economy.withdrawPlayer(player, XPBoostMain.boostCfg.getDouble(boostId + ".cost"));
+                } catch (BoostNotFoundException e) {
+                    e.printStackTrace();
                 }
-
-                if (XPBoostMain.boostCfg.contains(boostId + ".advanced")) {
-                    for (String pluginName : XPBoostMain.boostCfg.getConfigurationSection(boostId + ".advanced").getKeys(false)) {
-                        BoostOptions options = new BoostOptions(pluginName.toUpperCase());
-
-                        for (String option : XPBoostMain.boostCfg.getConfigurationSection(boostId + ".advanced." + pluginName).getKeys(false)) {
-                            if (option.equalsIgnoreCase("default")) {
-                                options.setEnabledByDefault(XPBoostMain.boostCfg
-                                        .getBoolean(boostId + ".advanced." + pluginName + "." + option));
-                            } else {
-                                options.getOptions().put(option.toUpperCase(), XPBoostMain.boostCfg
-                                        .getBoolean(boostId + ".advanced." + pluginName + "." + option));
-                            }
-                        }
-
-                        xpb.getAdvancedOptions().put(pluginName.toUpperCase(), options);
-                    }
-                }
-
             } else {
                 String message = XPBoostMain.getLang().getString("lang.buyfail");
                 message = message.replaceAll("%money%", XPBoostMain.boostCfg.getString(boostId + ".cost"));
@@ -76,5 +48,55 @@ public class BoostAPI {
         }
         
         return true;
+    }
+    
+    /**
+     * Force give player a specific boost
+     * 
+     * @param player
+     * @param boostId
+     * @param durationInSeconds Put null if you dont want to override defined duration from the boost's config
+     * @throws BoostNotFoundException
+     */
+    public void giveBoost(Player player, String boostId, Integer durationInSeconds) throws BoostNotFoundException {
+        if (!XPBoostMain.boostCfg.contains(boostId)) {
+            throw new BoostNotFoundException();
+        }
+        
+        int duration = durationInSeconds != null ? durationInSeconds : XPBoostMain.boostCfg.getInt(boostId + ".time");
+        double boost = XPBoostMain.boostCfg.getDouble(boostId + ".boost");
+
+        String message = XPBoostMain.getLang().getString("lang.xpbuy").replaceAll("%time%", duration + "")
+                .replaceAll("%money%", XPBoostMain.boostCfg.getString(boostId + ".cost"))
+                .replaceAll("%boost%", String.valueOf(boost));
+        MainAPI.sendMessage(message, player);
+
+        XPBoost xpb = XPBoostAPI.setPlayerBoost(player.getUniqueId(), boost, duration);
+
+        if (XPBoostMain.boostCfg.contains(boostId + ".behaviour")) {
+            for (String cond : XPBoostMain.boostCfg.getConfigurationSection(boostId + ".behaviour")
+                    .getKeys(false)) {
+                xpb.putCondition(Condition.valueOf(cond.toUpperCase()),
+                        XPBoostMain.boostCfg.getBoolean(boostId + ".behaviour." + cond));
+            }
+        }
+
+        if (XPBoostMain.boostCfg.contains(boostId + ".advanced")) {
+            for (String pluginName : XPBoostMain.boostCfg.getConfigurationSection(boostId + ".advanced").getKeys(false)) {
+                BoostOptions options = new BoostOptions(pluginName.toUpperCase());
+
+                for (String option : XPBoostMain.boostCfg.getConfigurationSection(boostId + ".advanced." + pluginName).getKeys(false)) {
+                    if (option.equalsIgnoreCase("default")) {
+                        options.setEnabledByDefault(XPBoostMain.boostCfg
+                                .getBoolean(boostId + ".advanced." + pluginName + "." + option));
+                    } else {
+                        options.getOptions().put(option.toUpperCase(), XPBoostMain.boostCfg
+                                .getBoolean(boostId + ".advanced." + pluginName + "." + option));
+                    }
+                }
+
+                xpb.getAdvancedOptions().put(pluginName.toUpperCase(), options);
+            }
+        }
     }
 }