+++ src/cz/dubcat/xpboost/api/messages/MessageLocation.java
@@ -1,0 +1,5 @@
+package cz.dubcat.xpboost.api.messages;
+
+public enum MessageLocation {
+    CHAT, ACTIONBAR
+}