+++ src/cz/dubcat/xpboost/exceptions/BoostNotFoundException.java
@@ -1,0 +1,5 @@
+package cz.dubcat.xpboost.exceptions;
+
+public class BoostNotFoundException extends Exception {
+    private static final long serialVersionUID = 2889752000839383924L;
+}