+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Change Log
+## [3.2.0] - 27/07/2019
+### Added
+- Vietnamese language (VN) thanks to @fecxica
+- MyPets support
+- command `/xpb giveDefinedBoost <player/all> <boostId> [durationInSeconds]` - gives a player (or everyone) some predefined boost _(boost does not have to be enabled in the config)_
+- ability to choose where you want your messages to appear in the `CHAT` or `ACTIONBAR` for `experienceGainedMessagesOptions` and `activeBoostReminderOptions` settings
+
+### Changed
+
+- renamed some configuration keys for better readability **(you may want to regenerate your config file)**
 
 ## [3.1.0] - 3/06/2019